+++ R/utils.R
@@ -966,7 +966,7 @@ getFCSFileName <- function(ff) {
 #' rawDataDir <-
 #'     system.file("extdata", package = "CytoPipeline")
 #' sampleFiles <-
-#'     file.path(rawDataDir, list.files(rawDataDir, pattern = "sample_"))
+#'     file.path(rawDataDir, list.files(rawDataDir, pattern = "Donor"))
 #' 
 #' truncateMaxRange <- FALSE
 #' minLimit <- NULL